Move rrf unit tests to rrf.py
